setup_mock_db.py
import os
from supabase import create_client
from dotenv import load_dotenv

# Provide a raw DB setup if we don't have RLS service role

def execute_raw_sql():
    # If the user has SQLAlchemy and psycopg2 installed we can connect bypassing RLS completely.
    # Otherwise we just instruct the user to run the curl command or skip the DB injection
    pass

if __name__ == "__main__":
    print("Please use the Supabase Dashboard to insert a user and an exercise.")
