requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
supabase==2.5.1
pytest==8.2.0
httpx==0.27.0
pydantic==2.7.1
python-dotenv==1.0.1
requests==2.32.3
Pillow==10.3.0


